lib/backend/ingestion/embeddings/requirements.txt
# Combined requirements from embeddings, langchain-layer, and ws-authorizer-layer
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amazon-textract-textractor==1.7.11
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
boto3==1.37.28
botocore==1.37.28
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cryptography==44.0.0
dataclasses-json==0.6.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-aws==0.2.18
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
langsmith==0.3.24
lxml==5.2.2
marshmallow==3.26.1
multidict==6.3.2
mypy-extensions==1.0.0
numpy>=1.21,<2
opensearch-py==2.8.0
orjson==3.10.16
packaging==24.2
pdfplumber==0.11.0
Pillow==10.4.0
pg8000==1.30.5
pgvector==0.2.5
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.8.0
pypdf==5.4.0
python-docx==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-aws4auth==1.2.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
aws-lambda-powertools==2.36.0
aws-xray-sdk==2.12.1
ffmpeg-python==0.2.0
markdown==3.6